[#619] refactor: PlaygroundPart enum 클래스명을 SoptPart 로 변경 - 인증중앙화로 인해 파트 정보를 받아와서 사용하는 곳이 playground에서 platform으로 변경됨. 추후 이런 류의 수정이 발생하더라도 공통적으로 사용할 수 있도록 네이밍을 SoptPart 로 변경
