+++ src/main/java/org/sopt/app/application/app_service/SoptampBadgeManager.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.sopt.app.application.app_service.dto.AppServiceBadgeInfo;
 import org.sopt.app.application.soptamp.SoptampUserService;
 import org.sopt.app.domain.enums.Part;
-import org.sopt.app.domain.enums.PlaygroundPart;
+import org.sopt.app.domain.enums.SoptPart;
 import org.sopt.app.facade.RankFacade;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
@@ -19,7 +19,7 @@ public class SoptampBadgeManager implements AppServiceBadgeManager {
 
     @Override
     public AppServiceBadgeInfo acquireAppServiceBadgeInfo(final Long userId) {
-        Part part = PlaygroundPart.toPart(soptampUserService.getSoptampUserInfo(userId).getPart());
+        Part part = SoptPart.toPart(soptampUserService.getSoptampUserInfo(userId).getPart());
         if(part == null) {
             return AppServiceBadgeInfo.createWithAllDisabled();
         }
