+++ src/main/java/org/sopt/app/presentation/user/UserResponse.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 import org.sopt.app.application.app_service.dto.AppServiceInfo;
 import org.sopt.app.application.playground.dto.PlaygroundProfileInfo.PlaygroundProfile;
-import org.sopt.app.domain.enums.PlaygroundPart;
+import org.sopt.app.domain.enums.SoptPart;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class UserResponse {
@@ -175,7 +175,7 @@ public class UserResponse {
                     .profileImage(playgroundProfile.getProfileImage() != null ? playgroundProfile.getProfileImage() : "")
                     .part(playgroundProfile.getAllActivities().stream()
                             .map(c -> c.getPlaygroundPart().getPartName())
-                            .filter(c -> !c.equals(PlaygroundPart.NONE.getPartName()))
+                            .filter(c -> !c.equals(SoptPart.NONE.getPartName()))
                             .collect(Collectors.joining("/")))
                     .profileMessage(playgroundProfile.getIntroduction() != null ? playgroundProfile.getIntroduction() : "")
                     .during(during != null ? during + "개월" : "")
